Go though the failing tests with the new test files

+++ src/modules/logs/content/log_event_content/clear_impound_event.rs
@@ -7,7 +7,7 @@ pub struct ClearImpoundEvent {
     pub ship_type: ShipType,
 
     #[serde(rename = "ShipID")]
-    pub ship_id: u8,
+    pub ship_id: u64,
 
     #[serde(rename = "ShipMarketID")]
     pub ship_market_id: u64,
